+++ tests/integration/test_keyvault_complete_replication.py
@@ -18,13 +18,14 @@ Test coverage:
 
 import json
 from typing import Any, Dict, List
-from unittest.mock import Mock
+from unittest.mock import AsyncMock, Mock, patch
 
 import pytest
 
 from src.config_manager import AzureTenantGrapherConfig
 from src.iac.emitters.terraform.context import EmitterContext
 from src.iac.emitters.terraform.handlers import ensure_handlers_registered
+from src.services.azure_discovery_service import AzureDiscoveryService
 
 
 class TestKeyVaultCompleteReplication:
@@ -282,9 +283,7 @@ class TestKeyVaultCompleteReplication:
 
         # Verify private_connection_resource_id references Key Vault
         psc = config["private_service_connection"]
-        assert (
-            "${azurerm_key_vault.test_kv.id}" in psc["private_connection_resource_id"]
-        )
+        assert "${azurerm_key_vault.test_kv.id}" in psc["private_connection_resource_id"]
 
     @pytest.mark.asyncio
     async def test_neo4j_relationships_preserved_in_terraform(
